Implement 2 daily mission recording limit, 5 pts award, front camera selector, and dynamic appointment stats in Android app

+++ app/src/main/java/com/example/smileapp/models/User.kt
@@ -30,7 +30,7 @@ data class User(
     @JvmField var specialization: String? = null,
 
     @JvmField var points: Int = 0,
-    @JvmField var streak: Int = 0,
+    @JvmField var streak: Double = 0.0,
     @JvmField var stars: Int = 0,
 
     @SerialName("warning_note")
